% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxM.R
\name{boxM}
\alias{boxM}
\title{Box's M-test}
\usage{
boxM(Y, group, tol = 1e-10, fxn = cor, method = c("pearson",
  "kendall", "spearman"))
}
\arguments{
\item{Y}{response matrix}

\item{group}{factor defining groups}

\item{tol}{tolerance for eigen values}

\item{fxn}{define function.  Here default is cor to compare correlation structure.  Use cov to compare covariance structure like in  heplots::boxM}

\item{method}{specify which correlation method: "pearson", "kendall" or "spearman"}
}
\description{
Box's M-test
}
\examples{
data(iris)

boxM( iris[,1:4], iris[,5])

}
