% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sled.R
\name{boxM_permute}
\alias{boxM_permute}
\title{Box's M-test}
\usage{
boxM_permute(Y, group, nperm = 150, tol = 1e-10, fxn = cor,
  method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{Y}{response variable matrix}

\item{group}{a factor defining groups, number of entries must equal nrow(Y)}

\item{nperm}{number of permutations of group variable used to estimate degrees of freedom under the null}
}
\value{
list of p.value, test statistic, and df.approx estimated by permutation
}
\description{
boxM performs the Box's (1949) M-test for homogeneity of covariance matrices obtained from multivariate normal data according to one or more classification factors. The test compares  the product of the log determinants of the separate covariance  matrices to the log determinant of the pooled covariance matrix,    analogous to a likelihood ratio test. The test statistic uses a chi-square approximation. Uses permutations to estimate the degrees of freedom under the null
}
\examples{
data(iris)

boxM_permute(iris[, 1:4], iris[, "Species"])

}
\seealso{
heplots::boxM
}
