% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sled.R
\name{corrMatrix.test}
\alias{corrMatrix.test}
\title{Test difference between two correlation matricies}
\usage{
corrMatrix.test(Y, group, method = c("Box", "Box.permute",
  "Steiger.fisher", "Steiger", "Jennrich", "Factor", "Mann.Whitney",
  "Kruskal.Wallis", "Cai.max", "Chang.maxBoot", "LC.U.test", "WL.randProj",
  "Schott.Frob", "Delaneau"), method2 = c("pearson", "kendall",
  "spearman"))
}
\arguments{
\item{Y}{data matrix}

\item{group}{a factor defining groups}

\item{method}{Specify test: "Box", "Box.permute", "Steiger.fisher", "Steiger", "Jennrich", "Factor", "Mann.Whitney" "Kruskal.Wallis", "Cai.max", "Chang.maxBoot", "LC.U.test", "WL.randProj", "Schott.Frob", "Delaneau"}

\item{method2}{Specify type of correlation: "pearson", "kendall", "spearman"}
}
\description{
Test difference between two correlation matricies using one of 5 tests
}
