% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{countClusters}
\alias{countClusters}
\title{Count clusters on each chromosome}
\usage{
countClusters(treeListClusters)
}
\arguments{
\item{treeListClusters}{from createClusters()}
}
\value{
count number of clusters on each chromsome
}
\description{
Count clusters on each chromosome
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clustering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clutsers
treeListClusters = createClusters( treeList )

# Count clusters on each chromsome
countClusters( treeListClusters )

}
