% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{createClusters}
\alias{createClusters}
\title{Create cluster from list of hclust objects}
\usage{
createClusters(treeList, method = c("capushe", "bstick",
  "meanClusterSize"), meanClusterSize = 50, pct = 0.15)
}
\arguments{
\item{treeList}{list of hclust objects}

\item{method}{'capushe': slope heuristic. 'bstick': broken stick. 'meanClusterSize': create clusters based on target mean value.}

\item{meanClusterSize}{select target mean cluster size}

\item{pct}{minimum percentage of points for the plateau selection in capushe selection.}
}
\value{
Convert hierarchical clustering into discrete clusters based on selection criteria method
}
\description{
Create cluster from list of hclust objects
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clustering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clutsers
treeListClusters = createClusters( treeList )

# Plot correlations and clusters in region defind by query
query = GRanges('chr1', IRanges(0, 1000))

plotDecorate( treeList, treeListClusters, query)

}
