% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{createCorrelationMatrix}
\alias{createCorrelationMatrix}
\title{Create correlation matrix}
\usage{
createCorrelationMatrix(
  query,
  regionQuant,
  adjacentCount = 500,
  windowSize = 1e+06,
  method = "adjacent",
  method.corr = c("pearson", "spearman"),
  quiet = FALSE,
  setNANtoZero = FALSE
)
}
\arguments{
\item{query}{GRanges object of intervals to query}

\item{regionQuant}{normalized quantifications of regions in query.  Rows are features, like in limma}

\item{adjacentCount}{number of adjacent entries to compute correlation against}

\item{windowSize}{width of window in bp around each interval beyond which weight is zero}

\item{method}{'adjacent': compute corr on fixed count sliding window define by adjacentCount.  "distance": compute corr for entries within windowSize bp}

\item{method.corr}{specify which correlation method: "pearson" or "spearman"}

\item{quiet}{suppress messages}

\item{setNANtoZero}{replace NAN correlation values with a zero}
}
\value{
for peak i and j with distance d_{i,j},
M[i,j] = cor( vobj$E[i,], vobj$E[j,] )

return sparse symmatric matrix
}
\description{
Create correlation matrix based on correlation between pairs of peaks
}
\examples{

data('decorateData')

C = createCorrelationMatrix(simLocation, simData)

}
