% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\docType{data}
\name{decorateData}
\alias{decorateData}
\alias{simData}
\alias{simLocation}
\alias{metadata}
\title{Simulated data to show correlation clustering}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 448 rows and 1000 columns.

An object of class \code{GRanges} of length 448.

An object of class \code{data.frame} with 1000 rows and 1 columns.
}
\usage{
data(decorateData)

data(decorateData)

data(decorateData)
}
\description{
Simulated data to show correlation clustering

Simulated data to show correlation clustering. GRanges object indicating genomic position of data in rows of simData

Simulated disease status for differential analysis
}
\keyword{decorateData}
