% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_corr_tests.R
\name{delaneau.score}
\alias{delaneau.score}
\title{Score impact of each sample on correlation sturucture}
\usage{
delaneau.score(Y, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{Y}{data matrix with samples on rows and variables on columns}

\item{method}{specify which correlation method: "pearson", "kendall" or "spearman"}
}
\value{
score for each sample measure impact on correlation structure
}
\description{
Score impact of each sample on correlation sturucture.  Compute correlation using all samples (i.e. C), then compute correlation omitting sample i (i.e. Ci).  The score the sample i is based on the difference between C and Ci.
}
\examples{
# load iris data
data(iris)

# Evalaute score on each sample
delaneau.score( iris[,1:4] )

}
\seealso{
delaneau.test
}
