% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_corr_tests.R
\name{delaneau.test}
\alias{delaneau.test}
\title{Test association between correlation sturucture and variable}
\usage{
delaneau.test(Y, variable, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{Y}{data matrix with samples on rows and variables on columns}

\item{variable}{variable with number of entries must equal nrow(Y).  Can be discrete or continuous.}

\item{method}{specify which correlation method: "pearson", "kendall" or "spearman"}
}
\value{
score for each sample measure impact on correlation structure
}
\description{
Score impact of each sample on correlation sturucture and then peform test of association with variable using Kruskal-Wallis test
}
\examples{
# load iris data
data(iris)

# Evalaute score on each sample
delaneau.test( iris[,1:4], iris[,5] )

}
\seealso{
delaneau.score
}
