% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sled.R
\name{.evalDiffCorr}
\alias{.evalDiffCorr}
\title{Internal .evalDiffCorr}
\usage{
.evalDiffCorr(epiSignal, testVariable, gRanges, clustList,
  npermute = c(100, 10000, 1e+05), adj.beta = 0, rho = 0,
  sumabs.seq = 1, BPPARAM = SerialParam(), method = c("sLED",
  "medianCorr"))
}
\arguments{
\item{method}{"sLED" or "medianCorr"}
}
\description{
Internal .evalDiffCorr
}
