% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\docType{class}
\name{epiclustDiscreteListContain-class}
\alias{epiclustDiscreteListContain-class}
\title{Class epiclustDiscreteListContain}
\description{
Class \code{epiclustDiscreteListContain}: is a list containing epiclustDiscreteList objects
}
