% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{filterClusters}
\alias{filterClusters}
\title{Extract subset of clusters}
\usage{
filterClusters(treeListClusters, chroms, clusters)
}
\arguments{
\item{treeListClusters}{from createClusters()}

\item{chroms}{array of chromosome names for the clusters to retain}

\item{clusters}{array of cluster names for the clusters to retain}
}
\value{
epiclustDiscreteList of specified clusters
}
\description{
Extract subset of clusters based on entries in chroms and clusters
}
\details{
Each cluster has a unique identifier: c(chroms[i], clusters[i]).  Return epiclustDiscreteList object like treeListClusters but retaining on the specified clusters.
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clustering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList )

# Evaluate score for each cluster
clstScore = scoreClusters(treeList, treeListClusters )

# Retain clusters that pass this criteria
clustInclude = clstScore[clstScore$LEF > 0.2,]

# get retained clusters
treeListClusters_filter = filterClusters( treeListClusters, clustInclude$chrom, clustInclude$cluster)

}
