% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnsGenes.R
\name{get_exon_coords}
\alias{get_exon_coords}
\title{Get coordinates of exons}
\usage{
get_exon_coords(ensdb, query, biotypes = c("protein_coding"))
}
\arguments{
\item{ensdb}{ENSEMBL database object like EnsDb.Hsapiens.v86}

\item{query}{GRranges ofject of one interval.  "chr20" should be coded as "20"}

\item{biotypes}{gene biotypes to return}
}
\value{
GRanges object of exon locations
}
\description{
Get coordinates of exons from ENSEMBL database
}
\examples{
library(EnsDb.Hsapiens.v86)
library(GenomicRanges) 

# gene database
ensdb = EnsDb.Hsapiens.v86

# interval
query = GRanges("20", IRanges(62045027,62164563))

# get GRanges object of exon locations
get_exon_coords( ensdb, query)

}
