% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCorrelationScores.R
\name{getClusterNames}
\alias{getClusterNames}
\title{Get name of each cluster}
\usage{
getClusterNames(clustList)
}
\arguments{
\item{clustList}{list of cluster assignments}
}
\value{
array of cluster names
}
\description{
Get name of each cluster as parameter:chrom:cluster
}
\examples{
library(GenomicRanges)

# load data
data('decorateData')

# Evaluate hierarchical clustering
# adjacentCount is the number of adjacent peaks considered in correlation
treeList = runOrderedClusteringGenome( simData, simLocation)

# Choose cutoffs and return cluster
treeListClusters = createClusters( treeList, method = "meanClusterSize", meanClusterSize=c( 10, 20, 30, 40, 50) )

# Name of each cluster is parameter:chrom:cluster
getClusterNames( treeListClusters)

}
