% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCorrelationScores.R
\name{getClusterRanges}
\alias{getClusterRanges}
\title{Get genome coordinates for each cluster}
\usage{
getClusterRanges(gRanges, clustList, verbose = TRUE)
}
\arguments{
\item{gRanges}{GenomciRanges corresponding to the rows of epiSignal}

\item{clustList}{list of cluster assignments}

\item{verbose}{write messages to screen}
}
\value{
GRanges object
}
\description{
Get genome coordinates for each cluster as a GRanges object
}
\examples{
library(GenomicRanges)

# load data
data('decorateData')

# Evaluate hierarchical clustering
# adjacentCount is the number of adjacent peaks considered in correlation
treeList = runOrderedClusteringGenome( simData, simLocation)

# Choose cutoffs and return cluster
treeListClusters = createClusters( treeList, method = "meanClusterSize", meanClusterSize=c( 10, 20, 30, 40, 50) )

# Get start and end coordinates for each cluster
# cluster name is parameter:chrom:cluster
getClusterRanges( simLocation, treeListClusters)

}
