% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{getFeaturesInCluster}
\alias{getFeaturesInCluster}
\title{Get feature names in selected cluster}
\usage{
getFeaturesInCluster(treeListClusters, chrom, clustID, id)
}
\arguments{
\item{treeListClusters}{from createClusters()}

\item{chrom}{chromosome name of cluster}

\item{clustID}{cluster identifier}

\item{id}{clustering parameter identifier}
}
\value{
array of feature names
}
\description{
Get feature names in selected cluster given chrom and clusterid
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clustering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList, method='meanClusterSize', meanClusterSize = 50 )

# Find chromsome and cluster of peak_204
getFeaturesInCluster( treeListClusters, "chr20", 3, "50")

}
