% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{getFeaturesInClusterList}
\alias{getFeaturesInClusterList}
\title{Get feature names in selected cluster}
\usage{
getFeaturesInClusterList(treeListClusters, chrom, clustID, id)
}
\arguments{
\item{treeListClusters}{from createClusters()}

\item{chrom}{chromosome name of cluster}

\item{clustID}{cluster identifier}

\item{id}{clustering parameter identifier}
}
\value{
list of array of feature names.  Query with index i as returned in list index i
}
\description{
Get feature names in selected cluster given array of chrom and cluster ids
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clustering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList, method='meanClusterSize', meanClusterSize = 50 )

df_cluster = data.frame( chrom = c("chr20", "chr20"), 
  cluster = c(3,5), 
  id = c("50", "50"), stringsAsFactors=FALSE )

# Find features for the clusters in df_cluster
getFeaturesInClusterList( treeListClusters, chrom=df_cluster$chrom, clustID=df_cluster$cluster, id=df_cluster$id)

}
