% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\docType{methods}
\name{getSubset}
\alias{getSubset}
\alias{getSubset,epiclust,GRanges-method}
\alias{getSubset,epiclustList,GRanges-method}
\title{Extract subset of data points}
\usage{
getSubset(fit, query)

\S4method{getSubset}{epiclust,GRanges}(fit, query)

\S4method{getSubset}{epiclustList,GRanges}(fit, query)
}
\arguments{
\item{fit}{epiclust or epiclustList object}

\item{query}{GenomicRanges object of which data points to retain}
}
\value{
epiclust or epiclustList object
}
\description{
Extract subset of data points
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clustering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# extract subset of data after clustering 
res = getSubset( treeList, simLocation[1:10])


}
