% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateCorrDecay.R
\name{ggplot_by_sampling}
\alias{ggplot_by_sampling}
\title{Plot by subsampling in each bin}
\usage{
ggplot_by_sampling(x, y, N, nbins = 1000)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{N}{number of samples}

\item{nbins}{number of bins on the x-axis}
}
\description{
Plot by subsampling in each bin in x-axis
}
