% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{jaccard}
\alias{jaccard}
\title{Evaluate Jaccard index}
\usage{
jaccard(a, b)
}
\arguments{
\item{a}{set 1}

\item{b}{set 2}
}
\value{
Jaccard index
}
\description{
Evaluate Jaccard index
}
\examples{
a = 1:10
b = 5:15
jaccard(a,b)

}
