% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotClusterSegments}
\alias{plotClusterSegments}
\title{Plot cluster segments}
\usage{
plotClusterSegments(clusterValues)
}
\arguments{
\item{clusterValues}{array of names of cluster for each entry}
}
\value{
ggplot2 of cluster assignments
}
\description{
Plot bar of segments showing clusters
}
\examples{
plotClusterSegments(c(rep(1, 5), rep(2,2), rep(3, 4)))

}
