% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCompareCorr}
\alias{plotCompareCorr}
\title{Plot two correlation matrices together}
\usage{
plotCompareCorr(epiSignal, peakIDs, testVariable, cols, size = 5,
  absCorr = FALSE)
}
\arguments{
\item{epiSignal}{matrix or EList of epigentic signal.  Rows are features and columns are samples}

\item{peakIDs}{feature names to extract from rows of epiSignal}

\item{testVariable}{factor indicating two subsets of the samples to compare}

\item{cols}{array of color values}

\item{size}{size of text}

\item{absCorr}{show absolute correlations}
}
\value{
ggplot2 of combined correlation matrix
}
\description{
Combined plot of correlation matricies from cases and controls
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clsutering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList, method = "meanClusterSize", meanClusterSize=c( 10, 20) )

# Simulate variable to split dataset by
set.seed(1)
metadata = data.frame( Disease = factor(sample(0:1, ncol(simData), replace=TRUE)))

# get peak ID's from chr1, cluster 1
peakIDs = getFeaturesInCluster( treeListClusters, "chr1", 1, "10")

# plot comparison of correlation matrices for peaks in peakIDs
#  where data is subset by metadata$Disease
plotCompareCorr( simData, peakIDs, metadata$Disease) + ggtitle("chr1: cluster 1")

}
