% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCorrTriangle}
\alias{plotCorrTriangle}
\title{Plot triangle of correlation matrix}
\usage{
plotCorrTriangle(C, size = 1, stroke = 1.5, cols = c("blue", "white",
  "red"), absCorr = FALSE)
}
\arguments{
\item{C}{correlation matrix}

\item{size}{plotting argument to geom_point()}

\item{stroke}{plotting argument to geom_point()}

\item{cols}{array of two colors for gradient}

\item{absCorr}{show absolute correlations}
}
\value{
ggplot2 plot of correlation matrix
}
\description{
Plot lower triangle of correlation matrix
}
\details{
Adjust size and stroke of points in the plot to fix look of plot depending on dimensions
}
\examples{
N = 1000
p = 100
X = matrix(rnorm(N*p), N,p)
C = cor(X)
plotCorrTriangle( C )

}
