% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateCorrDecay.R
\name{plotDensityPoints}
\alias{plotDensityPoints}
\title{Plot density as color, add outlier points}
\usage{
plotDensityPoints(
  x,
  y,
  n = 100,
  outlierQuantile = 1e-05,
  densityExponent = 0.25
)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{n}{the number of equally spaced points at which the density is to be estimated.}

\item{outlierQuantile}{show points if density is less than this quantile}

\item{densityExponent}{color based on density^densityExponent}
}
\description{
Plot density as color, add outlier points
}
\examples{

x = rnorm(10000)
y = rnorm(10000)

plotDensityPoints( x, y)

}
