% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnsGenes.R
\name{plotEnsGenes}
\alias{plotEnsGenes}
\title{Plot ENSEMBL genes}
\usage{
plotEnsGenes(
  ensdb,
  minRange,
  maxRange,
  chromosome,
  plot_lines_distance = 0.03,
  vp = viewport(x = 0, y = 0.95, just = c("left", "top")),
  splice_variants = TRUE,
  non_coding = TRUE
)
}
\arguments{
\item{ensdb}{ENSEMBL database object like EnsDb.Hsapiens.v86}

\item{minRange}{start genome coordinate}

\item{maxRange}{end genome coordinate}

\item{chromosome}{chrom}

\item{plot_lines_distance}{veritcal distance between genes}

\item{vp}{viewport}

\item{splice_variants}{if TRUE, show multiple transcripts from the same gene}

\item{non_coding}{if TRUE, also show non-coding genes}
}
\value{
GRanges object of exon locations
}
\description{
Plot ENSEMBL genes in region
}
\examples{
library(EnsDb.Hsapiens.v86)
library(GenomicRanges) 
library(grid) 

# gene database
ensdb = EnsDb.Hsapiens.v86

# interval
query = GRanges("20", IRanges(62045027,62164563))

# plot genes
fig = plotEnsGenes( ensdb, start(query), end(query), seqnames(query))
grid.draw( fig )

}
