% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrPairs.R
\name{plotPairwiseScatter}
\alias{plotPairwiseScatter}
\title{Scatter plot of all pairs of variables stratified by test variable}
\usage{
plotPairwiseScatter(
  X,
  Y,
  testVariable,
  size = 1,
  cols = c("#00ff40", "deepskyblue"),
  axisLabels = c("show", "internal", "none"),
  title = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{X}{data.frame of variables}

\item{Y}{data.frame of variables}

\item{testVariable}{factor indicating two subsets of the samples to compare}

\item{size}{size of points}

\item{cols}{color to label samples in for two levels of testVariable}

\item{axisLabels}{either "show" to display axisLabels, "internal" for labels  in the diagonal plots, or "none" for no axis labels}

\item{title}{title}

\item{xlab}{xlab}

\item{ylab}{xlab}
}
\value{
ggplot2 of combined pairwise scatter plots
}
\description{
Make a scatterplot for each pair of variables in X and Y.  Dataset is divided in two based on value in testVariable
}
