% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrPairs.R
\name{plotScatterPairs}
\alias{plotScatterPairs}
\title{Scatter plot of all pairs of variables stratified by test variable}
\usage{
plotScatterPairs(epiSignal, peakIDs, testVariable, size = 1)
}
\arguments{
\item{epiSignal}{matrix or EList of epigentic signal.  Rows are features and columns are samples}

\item{peakIDs}{feature names to extract from rows of epiSignal}

\item{testVariable}{factor indicating two subsets of the samples to compare}

\item{size}{size of points}
}
\value{
ggplot2 of combined pairwise scatter plots
}
\description{
Scatter plot of all pairs of variables stratified by test variable
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clsutering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList, method = "meanClusterSize", meanClusterSize=c(10, 30) )

# get peak ID's from chr1, cluster 1
peakIDs = getFeaturesInCluster( treeListClusters, "chr20", 2, "30")

# plot comparison of correlation matrices for peaks in peakIDs
#  where data is subset by metadata$Disease
plotScatterPairs( simData, peakIDs, metadata$Disease) + ggtitle("chr20: cluster 1")

}
