% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{retainClusters}
\alias{retainClusters}
\title{Retain clusters by applying filter}
\usage{
retainClusters(clstScore, metric = "LEF", cutoff = 0.4)
}
\arguments{
\item{clstScore}{score each cluster using scoreClusters()}

\item{metric}{column of clstScore to use in filtering}

\item{cutoff}{retain cluster than exceed the cutoff for metric.  Can be array with one entry per entry in clstScore}
}
\value{
data.frame of chrom, clutser, id (the clustering parameter value), and the specified metric
}
\description{
Retain clusters by applying filter
}
\examples{
library(GenomicRanges)
library(BiocParallel)

data('decorateData')

# Evaluate hierarchical clustering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList )

# Evaluate score for each cluster
clstScore = scoreClusters(treeList, treeListClusters, BPPARAM = SerialParam() )

# Retain clusters that pass this criteria
clustInclude = retainClusters( clstScore, "LEF", 0.30 )

# Or filter by mean absolute correlation
# clustInclude = retainClusters( clstScore, "mean_abs_corr", 0.1 )

# get retained clusters
treeListClusters_filter = filterClusters( treeListClusters, clustInclude )

}
