% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sled.R
\name{runFastStat}
\alias{runFastStat}
\title{Test Difference in correlation using Mann-Whitney test}
\usage{
runFastStat(itObj)
}
\arguments{
\item{itObj}{iterator}
}
\description{
Test Difference in correlation using Mann-Whitney test
}
