% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{runOrderedClustering}
\alias{runOrderedClustering}
\title{Run hierarchical clustering preserving order}
\usage{
runOrderedClustering(X, gr, alpha = 0.5)
}
\arguments{
\item{X}{data matrix were *rows* are features in sequential order}

\item{gr}{GenomicRanges object corresponding to rows in X}

\item{alpha}{mixture parameter weigting correlations (alpha=0) versus chromosome distances (alpha=1)}
}
\value{
hclust tree
}
\description{
Run hierarchical clustering preserving sequential order of entries
}
\details{
Use hclustgeo in ClustGeo package to generate hierarchical clustering that preserves sequential order.

Chavent, et al. 2017. ClustGeo: an R package for hierarchical clustering with spatial constraints. arXiv:1707.03897v2 doi:10.1007/s00180-018-0791-1
}
\examples{
library(GenomicRanges)
library(EnsDb.Hsapiens.v86)

# load data
data('decorateData')

# load gene locations
ensdb = EnsDb.Hsapiens.v86

# Evaluate hierarchical clustering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList )

# Plot correlations and clusters in region defind by query
query = range(simLocation)

plotDecorate( ensdb, treeList, treeListClusters, simLocation, query)

}
