% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{scoreClusters}
\alias{scoreClusters}
\title{Compute scores for each cluster}
\usage{
scoreClusters(treeList, treeListClusters, BPPARAM = bpparam())
}
\arguments{
\item{treeList}{list of hclust objects}

\item{treeListClusters}{from createClusters()}

\item{BPPARAM}{parameters for parallel evaluation}
}
\value{
for all pairs of peaks within windowSize, report distance
}
\description{
For each cluster compute summary statistics for the cluster to measure how strong the correlation structure is.  Clusters with weak correlation structure can be dropped from downstream analysis.
}
\details{
For each cluster, extract the correlation matrix and return the mean absolute correlation; the 75th, 90th and 95th quantile absolute correlation, and LEF, the leading eigen-value fraction which is the fraction of variance explained by the leading eigen value of the matrix abs(C).
}
\examples{
library(GenomicRanges)
library(BiocParallel)

data('decorateData')

# Evaluate hierarchical clustering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList )

# Evaluate score for each cluster
clstScore = scoreClusters(treeList, treeListClusters, BPPARAM = SerialParam() )

}
