% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_corr_tests.R
\name{sle.score}
\alias{sle.score}
\title{Score impact of each sample on sparse leading eigen-value}
\usage{
sle.score(Y, method = c("pearson", "kendall", "spearman"), rho = 0.05,
  sumabs = 1)
}
\arguments{
\item{Y}{data matrix with samples on rows and variables on columns}

\item{method}{specify which correlation method: "pearson", "kendall" or "spearman"}

\item{rho}{a positive constant such that cor(Y) + diag(rep(rho,p)) is positive definite.}

\item{sumabs}{regularization paramter. Value of 1 gives no regularization, sumabs*sqrt(p) is the upperbound of the L_1 norm of v,controling the sparsity of solution. Must be between 1/sqrt(p) and 1.}
}
\value{
score for each sample measure impact on correlation structure
}
\description{
Score impact of each sample on sparse leading eigen-value.  Compute correlation using all samples (i.e. C), then compute correlation omitting sample i (i.e. Ci).  The score the sample i is based on sparse leading eigen-value of the diffrence between C and Ci.
}
\examples{
# load iris data
data(iris)

# Evalaute score on each sample
sle.score( iris[,1:4] )

}
\seealso{
sle.test
}
