% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_corr_tests.R
\name{sle.test}
\alias{sle.test}
\title{Test association between sparse leading eigen-value and variable}
\usage{
sle.test(Y, variable, method = c("pearson", "kendall", "spearman"),
  rho = 0, sumabs = 1)
}
\arguments{
\item{Y}{data matrix with samples on rows and variables on columns}

\item{variable}{variable with number of entries must equal nrow(Y).  Can be discrete or continuous.}

\item{method}{specify which correlation method: "pearson", "kendall" or "spearman"}

\item{rho}{a positive constant such that cor(Y) + diag(rep(rho,p)) is positive definite.}

\item{sumabs}{regularization paramter. Value of 1 gives no regularization, sumabs*sqrt(p) is the upperbound of the L_1 norm of v,controling the sparsity of solution. Must be between 1/sqrt(p) and 1.}
}
\value{
score for each sample measure impact on correlation structure
}
\description{
Score impact of each sample on sparse leading eigen-value and then peform test of association with variable using Kruskal-Wallis test
}
\examples{
# load iris data
data(iris)

# Evalaute score on each sample
sle.test( iris[,1:4], iris[,5] )

}
\seealso{
sle.score
}
