% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sled.R
\docType{class}
\name{sLEDresults-class}
\alias{sLEDresults-class}
\title{An S4 class that stores results of sLED analysis}
\description{
An S4 class that stores results of sLED analysis
}
\section{Slots}{

\describe{
\item{\code{.Data}}{list of sLED results}
}}

