% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\docType{methods}
\name{[,epiclustDiscreteListContain,ANY,ANY,ANY-method}
\alias{[,epiclustDiscreteListContain,ANY,ANY,ANY-method}
\title{Allow subsetting of epiclustDiscreteListContain}
\usage{
\S4method{[}{epiclustDiscreteListContain,ANY,ANY,ANY}(x, i, j, ...,
  drop = TRUE)
}
\arguments{
\item{x}{epiclustDiscreteListContain}

\item{i}{index 1}

\item{j}{index 2}

\item{...}{additional arguement}

\item{drop}{TRUE/FALSE}
}
\value{
subset of epiclustDiscreteListContain
}
\description{
Allow subsetting of epiclustDiscreteListContain
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clustering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList )

}
