% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sled.R
\name{summary,sLEDresults-method}
\alias{summary,sLEDresults-method}
\title{Summarize sLED analysis}
\usage{
\S4method{summary}{sLEDresults}(object)
}
\arguments{
\item{object}{sLEDresults}
}
\value{
data.frame
}
\description{
extract statistic and p-value for each cluster
}
