% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{whichCluster}
\alias{whichCluster}
\title{Find which cluster a peak is in}
\usage{
whichCluster(treeListClusters, feature_id, id = NULL)
}
\arguments{
\item{treeListClusters}{from createClusters()}

\item{feature_id}{name of query feature}

\item{id}{clustering parameter identifier}
}
\value{
data.frame of chromosome and cluster
}
\description{
Find which cluster a peak is in
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clustering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList, method='meanClusterSize', meanClusterSize = c(50, 100) )

# Find chromsome and cluster of peak_20
whichCluster( treeListClusters, 'peak_20')

# Find chromsome and cluster of peak_20 with clustering parameter 50 
#  corresponding to meanClusterSize 
whichCluster( treeListClusters, 'peak_20', "50")

}
