% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{collapseCluster}
\alias{collapseCluster}
\title{Collapse clusters based on jaccard index}
\usage{
collapseCluster(treeListClusters, featurePositions, jaccardCutoff = 0.9)
}
\arguments{
\item{treeListClusters}{from createClusters()}

\item{featurePositions}{GRanges object storing location of each feature}

\item{jaccardCutoff}{cutoff value for jaccard index}
}
\value{
subset of clusters in treeListClusters that passes cutoff
}
\description{
Collapse clusters if jaccard index between clusters excceds a cutoff
}
