% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corSubsetPairs}
\alias{corSubsetPairs}
\title{Compute correlations between pairs of features}
\usage{
corSubsetPairs(Y, idxi, idxj, silent = FALSE)
}
\arguments{
\item{Y}{matrix where columns are features}

\item{idxi}{indecies}

\item{idxj}{indecies}

\item{silent}{suppress messages}
}
\value{
Compute local correlations between for all k: cor(Y[,idxi[k]], Y[,idxj[k]])
}
\description{
Compute correlations between pairs of features given in idxi and idxj
}
\examples{
# Simulate simple dataset
N = 600
Y = matrix(rnorm(N*100), 100, N)

# select pairs to compute correlations between
i1 = sample.int(N, 200, replace=TRUE)
i2 = sample.int(N, 200, replace=TRUE)

# evaluate all piars
C = corSubsetPairs(Y, i1,i2)

# show value
C[i1[10], i2[10]]

# show values from evaluating this pair directly
cor(Y[,i1[10]], Y[,i2[10]])

}
