% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\docType{data}
\name{decorateData}
\alias{decorateData}
\alias{simData}
\alias{simLocation}
\title{Simulated data to show correlation clustering}
\format{An object of class \code{matrix} with 433 rows and 1000 columns.}
\usage{
data(decorateData)

data(decorateData)
}
\description{
Simulated data to show correlation clustering

Simulated data to show correlation clustering. GRanges object indicating genomic position of data in rows of simData
}
\keyword{decorateData}
