% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_corr_tests.R
\name{delaneau.test}
\alias{delaneau.test}
\title{Test association between correlation sturucture and variable}
\usage{
delaneau.test(Y, variable, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{Y}{data matrix with samples on rows and variables on columns}

\item{variable}{variable with number of entries must equal nrow(Y).  Can be discrete or continuous.}

\item{method}{specify which correlation method: "pearson", "kendall" or "spearman"}
}
\value{
list of p-value, estimate and method used
}
\description{
Score impact of each sample on correlation sturucture and then peform test of association with variable using Kruskal-Wallis test
}
\details{
The statistical test used depends on the variable specified.
if variable is factor with multiple levels, use Kruskal-Wallis test
if variable is factor with 2 levels, use Wilcoxon test
if variable is continuous, use Wilcoxon test
}
\examples{
# load iris data
data(iris)

# variable is factor with multiple levels
# use kruskal.test
delaneau.test( iris[,1:4], iris[,5] )

# variable is factor with 2 levels
# use wilcox.test
delaneau.test( iris[1:100,1:4], iris[1:100,5] )

# variable is continuous
# use cor.test with spearman
delaneau.test( iris[,1:4], iris[,1] )

}
\seealso{
delaneau.score sle.test
}
