% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sled.R
\name{.evalDiffCorr}
\alias{.evalDiffCorr}
\title{Internal .evalDiffCorr}
\usage{
.evalDiffCorr(epiSignal, testVariable, gRanges, clustList,
  npermute = c(100, 10000, 1e+05), adj.beta = 0, rho = 0,
  sumabs.seq = 1, BPPARAM = SerialParam(), method = c("sLED", "Box",
  "Box.permute", "Steiger.fisher", "Steiger", "Jennrich", "Factor",
  "Mann.Whitney", "Kruskal.Wallis"), method2 = c("pearson", "kendall",
  "spearman"))
}
\arguments{
\item{epiSignal}{matrix or EList of epigentic signal.  Rows are features and columns are samples}

\item{testVariable}{factor indicating two subsets of the samples to compare}

\item{gRanges}{GenomciRanges corresponding to the rows of epiSignal}

\item{clustList}{list of cluster assignments}

\item{npermute}{array of two entries with min and max number of permutations}

\item{adj.beta}{parameter for sLED}

\item{rho}{a large positive constant such that A(X)-A(Y)+diag(rep(rho,p)) is positive definite. Where p is the number of features}

\item{sumabs.seq}{sparsity parameter}

\item{BPPARAM}{parameters for parallel evaluation}

\item{method}{"sLED", "Box", "Box.permute", "Steiger.fisher", "Steiger", "Jennrich", "Factor", "Mann.Whitney" or "Kruskal.Wallis"}

\item{method2}{Specify type of correlation: "pearson", "kendall", "spearman"}
}
\value{
list of result by chromosome and clustList
}
\description{
Internal .evalDiffCorr
}
