% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{filterClusters}
\alias{filterClusters}
\title{Extract subset of clusters}
\usage{
filterClusters(treeListClusters, clustInclude)
}
\arguments{
\item{treeListClusters}{from createClusters()}

\item{clustInclude}{data.frame from retainClusters() indcating which clusters to include}
}
\value{
epiclustDiscreteList of specified clusters
}
\description{
Extract subset of clusters based on entries in chroms and clusters
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clustering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList )

# Evaluate score for each cluster
clstScore = scoreClusters(treeList, treeListClusters )

# Retain clusters that pass this criteria
clustInclude = retainClusters( clstScore, "LEF", 0.30 )

# get retained clusters
treeListClusters_filter = filterClusters( treeListClusters, clustInclude)

}
