% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{makeImageRect}
\alias{makeImageRect}
\title{Convert correlation matrix into triangle plot}
\usage{
makeImageRect(nrow, ncol, cols, name, byrow = TRUE, cutoff = 100)
}
\arguments{
\item{nrow}{nrow(C)}

\item{ncol}{ncol(C)}

\item{cols}{entries of C converted to color}

\item{name}{name of the plot}

\item{byrow}{process C by row}

\item{cutoff}{retain correlations up to cutoff features away}
}
\description{
Adapted from LDheatmap
}
