% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateCorrDecay.R
\name{plotCorrDecay}
\alias{plotCorrDecay}
\title{Plot correlation delay}
\usage{
plotCorrDecay(dfDist, n = 20000, nbins = 1000, method = c("R",
  "Rsq"), xlim = c(10, 1e+06))
}
\arguments{
\item{dfDist}{data.frame of distance and correlation from from evaluateCorrDecay()}

\item{n}{number of points to sample per bin}

\item{nbins}{number of bins on the x-axis}

\item{method}{on show either R or Rsq on y-axis}

\item{xlim}{min and max values for x-axis}
}
\description{
Plot correlation delay using subsampling
}
\details{
Plot correlation versus log10 distance.  Sample equal number of points for each bin along the x-axis.
}
\examples{
library(GenomicRanges)
library(ggplot2)

data('decorateData')

# Evaluate hierarchical clustering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Evaluate how correlation between features decays with distance
dfDist = evaluateCorrDecay( treeList, simLocation )

# make plot
plotCorrDecay( dfDist )

}
