% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDecorate}
\alias{plotDecorate}
\title{Plot decorate analysis}
\usage{
plotDecorate(treeList, treeListClusters, query, cols = c("white", "red"),
  plotTree = TRUE, windowWidth = 1000)
}
\arguments{
\item{treeList}{hierarchical clustering of each chromosome from runOrderedClusteringGenome()}

\item{treeListClusters}{assign regions to clusters after cutting tree with createClusters()}

\item{query}{GRanges object indiecating region to plot}

\item{plotTree}{show tree from hierachical clustering}

\item{windowWidth}{for each feature plot correlation with windowWidth adjacent features}
}
\value{
ggplot2 of cluster assignments and correlation between peaks
}
\description{
Plot decorate analysis for clusters and correlations
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clsutering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList )

# Plot correlations and clusters in region defind by query
query = GRanges('chr1', IRanges(0, 1000))

plotDecorate( treeList, treeListClusters, query)

}
