% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDecorate}
\alias{plotDecorate}
\title{Plot decorate analysis}
\usage{
plotDecorate(ensdb, treeList, treeListClusters, featurePositions, query,
  data, cols, showGenes = TRUE, splice_variants = FALSE,
  non_coding = FALSE, absCorr = FALSE, method.corr = c("pearson",
  "kendall", "spearman"))
}
\arguments{
\item{ensdb}{ENSEMBL database object like EnsDb.Hsapiens.v86}

\item{treeList}{hierarchical clustering of each chromosome from runOrderedClusteringGenome()}

\item{treeListClusters}{assign regions to clusters after cutting tree with createClusters()}

\item{featurePositions}{GRanges object storing location of each feature}

\item{query}{GRanges object indiecating region to plot}

\item{data}{data to compute correlations stratified by testVariable}

\item{cols}{array of color values}

\item{showGenes}{plot genes}

\item{splice_variants}{if TRUE, show multiple transcripts from the same gene}

\item{non_coding}{if TRUE, also show non-coding genes}

\item{absCorr}{show absolute correlations}

\item{method.corr}{if data is specified, compute correlation using: "pearson", "kendall", "spearman"}
}
\value{
ggplot2 of cluster assignments and correlation between peaks
}
\description{
Plot decorate analysis for clusters and correlations
}
\examples{
library(GenomicRanges)
library(EnsDb.Hsapiens.v86)

# load data
data('decorateData')

# load gene locations
ensdb = EnsDb.Hsapiens.v86

# Evaluate hierarchical clsutering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList, method='meanClusterSize', meanClusterSize=30) 

# Plot correlations and clusters in region defined by query
query = range(simLocation)

plotDecorate( ensdb, treeList, treeListClusters, simLocation, query)

}
