% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDecorate}
\alias{plotDecorate}
\title{Plot decorate analysis}
\usage{
plotDecorate(treeList, treeListClusters, query, size = 1, stroke = 1.5,
  cols = c("blue", "white", "red"), plotTree = TRUE)
}
\arguments{
\item{treeList}{hierarchical clustering of each chromosome from runOrderedClusteringGenome()}

\item{treeListClusters}{assign regions to clusters after cutting tree with createClusters()}

\item{query}{GRanges object indiecating region to plot}

\item{size}{plotting argument to geom_point() in correlation triangle}

\item{stroke}{plotting argument to geom_point() in correlation triangle}

\item{cols}{array of two colors for gradient in correlation triangle}

\item{plotTree}{show tree from hierachical clustering}
}
\value{
ggplot2 of cluster assignments and correlation between peaks
}
\description{
Plot decorate analysis for clusters and correlations
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clsutering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clutsers
treeListClusters = createClusters( treeList )

# Plot correlations and clusters in region defind by query
query = GRanges('chr1', IRanges(0, 1000))

plotDecorate( treeList, treeListClusters, query)

}
