% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDecorate}
\alias{plotDecorate}
\title{Plot decorate analysis}
\usage{
plotDecorate(treeList, treeListClusters, featurePositions, query,
  cols = c("lightyellow", "red"), showTree = TRUE, showGenes = FALSE)
}
\arguments{
\item{treeList}{hierarchical clustering of each chromosome from runOrderedClusteringGenome()}

\item{treeListClusters}{assign regions to clusters after cutting tree with createClusters()}

\item{featurePositions}{GRanges object storing location of each feature}

\item{query}{GRanges object indiecating region to plot}

\item{cols}{vector of two colors}

\item{showTree}{show tree from hierachical clustering}

\item{showGenes}{plot genes}
}
\value{
ggplot2 of cluster assignments and correlation between peaks
}
\description{
Plot decorate analysis for clusters and correlations
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clsutering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clusters
treeListClusters = createClusters( treeList )

# Plot correlations and clusters in region defind by query
query = range(simLocation)

plotDecorate( treeList, treeListClusters, simLocation, query)

}
