% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{retainClusters}
\alias{retainClusters}
\title{Retain clusters by applying filter}
\usage{
retainClusters(clstScore, metric = "LEF", cutoff = 0.4)
}
\arguments{
\item{clstScore}{score each cluster using scoreClusters()}

\item{metric}{column of clstScore to use in filtering}

\item{cutoff}{retain cluster than exceed the cutoff for metric}
}
\description{
Retain clusters by applying filter
}
