% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sled.R
\name{runFastStat}
\alias{runFastStat}
\title{Test difference in correlation using closed form tests}
\usage{
runFastStat(itObj, method)
}
\arguments{
\item{itObj}{iterator}

\item{method}{"Steiger.fisher", "Steiger", "Jennrich", "Factor" or "Mann.Whitney"}
}
\description{
Test difference in correlation using closed form tests
}
