% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_correlations.R
\name{whichCluster}
\alias{whichCluster}
\title{Find which cluster a peak is in}
\usage{
whichCluster(treeListClusters, id)
}
\arguments{
\item{treeListClusters}{from createClusters()}

\item{id}{name of query peak}
}
\value{
data.frame of chromosome and cluster
}
\description{
Find which cluster a peak is in
}
\examples{
library(GenomicRanges)

data('decorateData')

# Evaluate hierarchical clsutering
treeList = runOrderedClusteringGenome( simData, simLocation ) 

# Choose cutoffs and return clutsers
treeListClusters = createClusters( treeList )

# Find chromsome and cluster of peak_204
whichCluster( treeListClusters, 'peak_204')

}
